SPrintF: Clean up benign warnings
Update to VS2017, NUnitTestAdapter 2.2.0

IDE0017: Object initialization can be simplified.
IDE0018: Variable declaration can be inlined
IDE0019: Use pattern matching
IDE0028: Collection initialization can be simplified
IDE0041: Simplify null, `ReferenceEquals(object, null)` to `object is null`.
IDE0051: Remove unused private member
IDE0054: Use compound assignment
IDE0059: Unnecessary assignment of variable
IDE0067: Ensure objects are disposed
S101: Rename to match pascal casing rules
S125: Remove commented out code
S907: Remove use of goto (ignored)
S1116: Remove empty statement
S1121: Extract assignment from expression
S1144: Remove unused private field
S1168: Return empty collection instead of null
S1172: Remove unused method parameter
S1206: Overrides GetHashCode, so should override Equals
S1481: Remove unused local variable
S1764: Correct identical expressions on both sides of operator
S1854: Remove useless assignment to local variable
S1905: Remove unnecessary cast
S1939: Remove superfluous interfaces
S1939: Don't use 'int' as explicit underlying type for enum
S2178: Correct '|' to '||' and '&' to '&&' (false positive)
S2328: Refactor GetHashCode to not reference mutable fields
S2344: Rename enumeration to remove Flags suffix
S2346: Rename 'enum value' to 'None'
S2737: Add logic to catch clause or remove it
S3237: Use 'value'
S3889: Don't use Thread.Resume or Thread.Suspend
S3966: Dispose only once
S4070: Remove 'FlagsAttribute' from enum
S4143: Verify key was intended, value has already been set.

Use C# native types (int, uint, short, ushort).

Remove explicit usage of default base constructor.

Issue: HELIOS-1162

+++ test/Text/StringUtilities_SprintFPerformance.cs
@@ -12,11 +12,11 @@
         [Explicit("Performance Test")]
         public void SPrintF_Performance()
         {
-            PerformanceTest("SPrintF Double %e", 13, 10, 100000, () => { StringUtilities.SPrintF("%e", (double)123456.789); });
+            PerformanceTest("SPrintF Double %e", 13, 10, 100000, () => { StringUtilities.SPrintF("%e", 123456.789); });
             PerformanceTest("SPrintF Single %e", 13, 10, 100000, () => { StringUtilities.SPrintF("%e", (float)123456.789); });
-            PerformanceTest("SPrintF Double %f", 13, 10, 100000, () => { StringUtilities.SPrintF("%f", (double)123456.789); });
+            PerformanceTest("SPrintF Double %f", 13, 10, 100000, () => { StringUtilities.SPrintF("%f", 123456.789); });
             PerformanceTest("SPrintF Single %f", 13, 10, 100000, () => { StringUtilities.SPrintF("%f", (float)123456.789); });
-            PerformanceTest("SPrintF Double %g", 13, 10, 100000, () => { StringUtilities.SPrintF("%g", (double)123456.789); });
+            PerformanceTest("SPrintF Double %g", 13, 10, 100000, () => { StringUtilities.SPrintF("%g", 123456.789); });
             PerformanceTest("SPrintF Single %g", 13, 10, 100000, () => { StringUtilities.SPrintF("%g", (float)123456.789); });
             PerformanceTest("SPrintF Integer %d", 13, 10, 100000, () => { StringUtilities.SPrintF("%d", 16384); });
             PerformanceTest("SPrintF Unsigned %u", 13, 10, 100000, () => { StringUtilities.SPrintF("%u", 16384); });
@@ -51,8 +51,8 @@
         [Explicit("Performance Test")]
         public void SPrintF_PerformanceExponentDouble()
         {
-            PerformanceTest("SPrintF Double %e", 13, 10, 100000, () => { StringUtilities.SPrintF("%e", (double)123456.789); });
-            PerformanceTest(".NET Double {E}  ", 13, 10, 100000, () => { String.Format("{0:E}", (double)123456.789); });
+            PerformanceTest("SPrintF Double %e", 13, 10, 100000, () => { StringUtilities.SPrintF("%e", 123456.789); });
+            PerformanceTest(".NET Double {E}  ", 13, 10, 100000, () => { String.Format("{0:E}", 123456.789); });
 
             Assert.Pass();  // S2699: Performance test, user should check results.
         }
@@ -73,8 +73,8 @@
         [Explicit("Performance Test")]
         public void SPrintF_PerformanceFixedDouble()
         {
-            PerformanceTest("SPrintF Double %f", 13, 10, 100000, () => { StringUtilities.SPrintF("%f", (double)123456.789); });
-            PerformanceTest(".NET Double {F}  ", 13, 10, 100000, () => { String.Format("{0:F}", (double)123456.789); });
+            PerformanceTest("SPrintF Double %f", 13, 10, 100000, () => { StringUtilities.SPrintF("%f", 123456.789); });
+            PerformanceTest(".NET Double {F}  ", 13, 10, 100000, () => { String.Format("{0:F}", 123456.789); });
 
             Assert.Pass();  // S2699: Performance test, user should check results.
         }
@@ -95,8 +95,8 @@
         [Explicit("Performance Test")]
         public void SPrintF_PerformanceGeneralDouble()
         {
-            PerformanceTest("SPrintF Double %g", 13, 10, 100000, () => { StringUtilities.SPrintF("%g", (double)123456.789); });
-            PerformanceTest(".NET Double {G}  ", 13, 10, 100000, () => { String.Format("{0:G}", (double)123456.789); });
+            PerformanceTest("SPrintF Double %g", 13, 10, 100000, () => { StringUtilities.SPrintF("%g", 123456.789); });
+            PerformanceTest(".NET Double {G}  ", 13, 10, 100000, () => { String.Format("{0:G}", 123456.789); });
 
             Assert.Pass();  // S2699: Performance test, user should check results.
         }
